Factor out common test trees.
