+++ src/lib/matchers.rs
@@ -261,42 +261,7 @@ pub fn has_same_type<T: Clone + Debug + Eq>(n1: &NodeId<FromNodeId>,
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    fn create_mult_arena() -> Arena<&'static str, FromNodeId> {
-        let mut arena = Arena::new();
-        let root = arena.new_node("Expr", String::from("+"), None, None, None, None);
-        let n1 = arena.new_node("INT", String::from("1"), None, None, None, None);
-        n1.make_child_of(root, &mut arena).unwrap();
-        let n2 = arena.new_node("Expr", String::from("*"), None, None, None, None);
-        n2.make_child_of(root, &mut arena).unwrap();
-        let n3 = arena.new_node("INT", String::from("3"), None, None, None, None);
-        n3.make_child_of(n2, &mut arena).unwrap();
-        let n4 = arena.new_node("INT", String::from("4"), None, None, None, None);
-        n4.make_child_of(n2, &mut arena).unwrap();
-        let format1 = "\"Expr\" +
-  \"INT\" 1
-  \"Expr\" *
-    \"INT\" 3
-    \"INT\" 4
-";
-        assert_eq!(format1, format!("{:?}", arena));
-        arena
-    }
-
-    fn create_plus_arena() -> Arena<&'static str, FromNodeId> {
-        let mut arena = Arena::new();
-        let root = arena.new_node("Expr", String::from("+"), None, None, None, None);
-        let n1 = arena.new_node("INT", String::from("3"), None, None, None, None);
-        n1.make_child_of(root, &mut arena).unwrap();
-        let n2 = arena.new_node("INT", String::from("4"), None, None, None, None);
-        n2.make_child_of(root, &mut arena).unwrap();
-        let format1 = "\"Expr\" +
-  \"INT\" 3
-  \"INT\" 4
-";
-        assert_eq!(format1, format!("{:?}", arena));
-        arena
-    }
+    use test_common::{create_mult_arena, create_plus_arena};
 
     #[test]
     fn is_isomorphic() {
